requirements.txt
pytest~=7.1.2
pandas~=1.4.3
pyserial~=3.5
nidaqmx~=0.6.3
PyVISA~=1.12.0
future~=0.18.2
pandastable~=0.13.0
Pillow~=9.1.1